Split up datahub.company.admin into multiple modules
